+++ src/weapon.h
@@ -24,19 +24,19 @@ typedef enum { WEAPON_LAYER_BG, WEAPON_LAYER_FG } WeaponLayer;
 /*
  * addition
  */
-void weapon_add( const Outfit* outfit, const double T,
+void weapon_add(const Outfit* outfit, const double T,
       const double dir, const Vector2d* pos, const Vector2d* vel,
-      const Pilot *parent, const unsigned int target, double time );
+      const Pilot *parent, const unsigned long target, double time);
 
 
 /*
  * Beam weapons.
  */
-unsigned int beam_start( const Outfit* outfit,
+unsigned int beam_start(const Outfit* outfit,
       const double dir, const Vector2d* pos, const Vector2d* vel,
-      const Pilot *parent, const unsigned int target,
-      PilotOutfitSlot *mount );
-void beam_end( const unsigned int parent, unsigned int beam );
+      const Pilot *parent, const unsigned long target,
+      PilotOutfitSlot *mount);
+void beam_end(const unsigned long parent, unsigned int beam);
 
 
 /*
